+++ src/main/kotlin/utils/Grid.kt
@@ -7,7 +7,7 @@ typealias Grid<Tile> = Map<Point, Tile>
 fun <Tile> BufferedReader.toGrid(tileMapper: (Char) -> Tile): Grid<Tile> = this.lineSequence()
     .flatMapIndexed { y, line ->
         line.mapIndexed { x, char ->
-            Point(x, y) to tileMapper(char)
+            Point(x.toLong(), y.toLong()) to tileMapper(char)
         }
     }.toMap()
